requirements.txt
flask==3.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
mysql-connector-python==8.2.0
pandas=2.1.3
openai=1.3.4